pyproject.toml
[tool.poetry]
authors = ["alexisthethe <alexisthethe@gmail.com>"]
description = "FastAPI project example"
name = "quickrun-fastapi"
version = "0.1.0"

[tool.poetry.dependencies]
SQLAlchemy = "^1.4.34"
fastapi = "^0.75.1"
passlib = "^1.7.4"
python = "^3.9"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
uvicorn = "^0.17.6"

[tool.poetry.dev-dependencies]
PyYAML = "^6.0"
black = "^22.3.0"
flake8 = "^4.0.1"
poethepoet = "^0.13.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
ipython = "^8.2.0"

[tool.poe.tasks.start-api]
envfile = ".env.dev"
sequence = [
  {cmd = "uvicorn app.main:app --port 8000 --reload"},
]

[tool.poe.tasks]
docker-build = [
  {cmd = "poetry export -f requirements.txt --output requirements.txt"},
  {cmd = "docker build -t quickrun-fastapi ."},
]
docker-run = "docker run -ti --rm -p 8000:80 --env-file .env.dev quickrun-fastapi"
lint = "flake8 app tests"
spec = {script = "app.commands:generate_spec_yaml()"}
test = "pytest -v --cov=./app --cov-report=term-missing tests"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
